modify

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: okamototakeshi <okamototakeshi@student.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/09 00:29:34 by okamototake       #+#    #+#             */
-/*   Updated: 2022/08/11 08:49:03 by okamototake      ###   ########.fr       */
+/*   Updated: 2022/08/11 10:05:15 by okamototake      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@ size_t			ft_strlen(const char *s);
 size_t			ft_put_c(const char c);
 size_t			ft_put_s(const char *str);
 size_t			ft_put_d_i(int num);
-size_t			ft_put_p(unsigned long int num);
+size_t			ft_put_p(uintptr_t num);
 size_t			ft_put_u(unsigned int num);
 size_t			ft_put_x(unsigned int num, bool sign);
 int				ft_printf(const char *fmt, ...);
